#450
Refactor and file structure change on functions having to do with QR code feature
