+++ src/routes/[link=link].svg.ts
@@ -1,19 +1,9 @@
-import QRCode from 'qrcode';
 import type { RequestEvent, RequestHandlerOutput } from '@sveltejs/kit/types/internal';
 import { links } from '$lib/constants/links';
-import { clean } from '$lib/constants/utils';
-
-function genQRCodeSvg(url: string): Promise<string> {
-  return new Promise((resolve, reject) => {
-    QRCode.toString(url, { type: 'svg' }, (err, svg) => {
-      if (err) reject(err);
-      resolve(svg);
-    });
-  });
-}
+import { normalizeLinkName, genQRCodeSvg } from '$lib/qr/utils';
 
 export async function get(event: RequestEvent<{ link: string }>): Promise<RequestHandlerOutput> {
-  const destination = links[clean(event.params.link)];
+  const destination = links[normalizeLinkName(event.params.link)];
   const payload = await genQRCodeSvg(destination);
 
   return new Response(payload, {
